Image: Use s32 instead of long
long is 64-bit on systems using the ILP64 and LP64 data model (with
LP64 being used by most Unix and Unix-based systems). However on systems
using the LLP64 data model (Windows), long is 32-bit. To prevent
portability problems in the future, we just use a fixed-width type here.

+++ Image.cpp
@@ -80,7 +80,7 @@ bool CImage::Init(CArcFile* archive, void* bmp, const YCString& file_name)
 /// @param pallet_size Palette size
 /// @param file_name   File name
 ///
-bool CImage::Init(CArcFile* archive, long width, long height, u16 bpp, const void* pallet, size_t pallet_size, const YCString& file_name)
+bool CImage::Init(CArcFile* archive, s32 width, s32 height, u16 bpp, const void* pallet, size_t pallet_size, const YCString& file_name)
 {
 	m_archive = archive;
 	m_option = archive->GetOpt();
