+++ Image/ImageBase.h
@@ -25,7 +25,7 @@ public:
 	CImageBase();
 	virtual ~CImageBase();
 
-	bool Init(CArcFile* archive, long width, long height, u16 bpp, const void* pallet = nullptr, size_t pallet_size = 1024, const YCString& file_name = _T(""));
+	bool Init(CArcFile* archive, s32 width, s32 height, u16 bpp, const void* pallet = nullptr, size_t pallet_size = 1024, const YCString& file_name = _T(""));
 
 	void SetValidityOfAlphaBlend(bool validity_of_alpha_blend);
 	bool GetValidityOfAlphaBlend() const;
@@ -33,9 +33,9 @@ public:
 	void SetBackColorWhenAlphaBlend(u32 background_color);
 
 	bool Compress(LPCTSTR dst_path, const void* bmp, size_t bmp_size);
-	bool Compress(LPCTSTR dst_path, const void* dib, size_t dib_size, const void* pallet, size_t pallet_size, u16 pallet_bpp, long width, long height, u16 bpp);
+	bool Compress(LPCTSTR dst_path, const void* dib, size_t dib_size, const void* pallet, size_t pallet_size, u16 pallet_bpp, s32 width, s32 height, u16 bpp);
 	bool Compress(void* dst, size_t dst_size, const void* bmp, size_t bmp_size);
-	bool Compress(void* dst, size_t dst_size, const void* dib, size_t dib_size, const void* pallet, size_t pallet_size, u16 pallet_bpp, long width, long height, u16 bpp);
+	bool Compress(void* dst, size_t dst_size, const void* dib, size_t dib_size, const void* pallet, size_t pallet_size, u16 pallet_bpp, s32 width, s32 height, u16 bpp);
 
 	bool ComposeBGRA(void* dst, const void* buffer, size_t buffer_size);
 	void WriteCompoBGRA(const void* buffer, size_t buffer_size, bool progress = true);
@@ -51,7 +51,7 @@ public:
 	void WriteFinish();
 
 protected:
-	static long CalculatePitch(long width, u16 bpp);
+	static s32 CalculatePitch(s32 width, u16 bpp);
 
 	bool IsRequireAlphaBlend() const;
 
@@ -75,12 +75,12 @@ protected:
 	bool   m_alpha_blend_requirement = false;
 	bool   m_output_dummy_from_buffer = false;
 
-	long   m_width = 0;
-	long   m_height = 0;
-	long   m_line = 0;
-	long   m_pitch = 0;
-	long   m_line_with_alpha = 0;
-	long   m_pitch_with_alpha = 0;
+	s32    m_width = 0;
+	s32    m_height = 0;
+	s32    m_line = 0;
+	s32    m_pitch = 0;
+	s32    m_line_with_alpha = 0;
+	s32    m_pitch_with_alpha = 0;
 	u16    m_bpp = 0;
 	u16    m_bpp_with_alpha = 0;
 	UPixel m_background_color_when_alpha_blending;
